Merge pull request #4173 from Penguinwizzard:arm64_inlining

The biggest thing that needed fixing up was the loading of inlinee
call info data; I implemented approximately the same thing that we
do on arm, minus doing it a little earlier (when encoding, instead
of in a relocation step). The most notable change is that I had to
make LabelInstr offsets grow to uintptr_t (from uint32); this will
not take any more memory, due to already being unioned with a byte
pointer, but has impact outside of ARM64.

The other part is the definition of two new instructions that have
a move with a known shift amount (movz_shift and movk_shift); this
is to ease definition of these fixed data moves, and may be useful
elsewhere.

+++ lib/Backend/arm64/LowerMD.cpp
@@ -7613,7 +7613,7 @@ LowererMD::FinalLower()
 
                 if (branchInstr->GetTarget() && !LowererMD::IsUnconditionalBranch(branchInstr)) //Ignore BX register based branches & B
                 {
-                    uint32 targetOffset = branchInstr->GetTarget()->GetOffset();
+                    uint32 targetOffset = (uint32)branchInstr->GetTarget()->GetOffset();
 
                     if (targetOffset != 0)
                     {
